🚚 Move Overview-component to new folder
